Test refactoring (#383)
