+++ src/cwe_checker_lib/src/analysis/pointer_inference/object/tests.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::intermediate_representation::Variable;
+use crate::{intermediate_representation::*, variable};
 use std::collections::BTreeMap;
 
 fn new_abstract_object() -> AbstractObject {
@@ -23,7 +23,7 @@ fn bv(number: i64) -> ValueDomain {
 fn new_id(tid: &str, reg_name: &str) -> AbstractIdentifier {
     AbstractIdentifier::new(
         Tid::new(tid),
-        AbstractLocation::Register(Variable::mock(reg_name, ByteSize::new(8))),
+        AbstractLocation::Register(variable!(format!("{reg_name}:8"))),
     )
 }
