+++ src/cwe_checker_lib/src/intermediate_representation/macros/mod.rs
@@ -189,10 +189,10 @@ pub mod parsing {
     #[allow(dead_code)]
     pub fn parse_expr<S: AsRef<str>>(str: S) -> Expression {
         let set = RegexSet::new([
-            r"^[[:alnum:]&&[^0-9]]{1}[[:alnum:]&&[^x]]?[[:alnum:]]*:[0-9]{1,2}$", // Variable
-            r"^((0x(-)?[[:alnum:]]+)|^(-)?([0-9])+)+:[0-9]+$",                    // Constant
-            r"^[^\+]*\+{1}[^\+]*$",                                               // BinOp (IntAdd)
-            r"^[[:ascii:]]+ \-{1} [[:ascii:]]+$",                                 // BinOp (IntSub)
+            r"^[[:alnum:]&&[^0-9]]{1}[[:alnum:]&&[^x]]?[[:alnum:]_]*:[0-9]{1,2}$", // Variable
+            r"^((0x(-)?[[:alnum:]]+)|^(-)?([0-9])+)+:[0-9]+$",                     // Constant
+            r"^[^\+]*\+{1}[^\+]*$",                                                // BinOp (IntAdd)
+            r"^[[:ascii:]]+ \-{1} [[:ascii:]]+$",                                  // BinOp (IntSub)
             r"^-\([[:ascii:]]*\)$",  // UnOp (IntNegate)
             r"^¬\([[:ascii:]]*\)$", // UnOp (BoolNegate)
         ])
