+++ src/cwe_checker_lib/src/intermediate_representation/program.rs
@@ -49,75 +49,3 @@ impl Program {
         None
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::intermediate_representation::{CallingConvention, Datatype};
-
-    use super::*;
-
-    impl Program {
-        fn add_extern_symbols_to_program(a: Vec<(Tid, ExternSymbol)>) -> Program {
-            Program {
-                subs: BTreeMap::new(),
-                extern_symbols: BTreeMap::from_iter(a),
-                entry_points: BTreeSet::new(),
-                address_base_offset: 0x1000u64,
-            }
-        }
-        /// Returns Program with malloc, free and other_function
-        pub fn mock_x64() -> Program {
-            let malloc = ExternSymbol::create_extern_symbol(
-                "malloc",
-                CallingConvention::mock_x64(),
-                Some(Datatype::Integer),
-                Some(Datatype::Pointer),
-            );
-            let free = ExternSymbol::create_extern_symbol(
-                "free",
-                CallingConvention::mock_x64(),
-                Some(Datatype::Pointer),
-                None,
-            );
-            let other_function = ExternSymbol::create_extern_symbol(
-                "other_function",
-                CallingConvention::mock_x64(),
-                None,
-                None,
-            );
-
-            Program::add_extern_symbols_to_program(vec![
-                (malloc.tid.clone(), malloc),
-                (free.tid.clone(), free),
-                (other_function.tid.clone(), other_function),
-            ])
-        }
-        /// Returns Program with malloc, free and other_function
-        pub fn mock_arm32() -> Program {
-            let malloc = ExternSymbol::create_extern_symbol(
-                "malloc",
-                CallingConvention::mock_arm32(),
-                Some(Datatype::Integer),
-                Some(Datatype::Pointer),
-            );
-            let free = ExternSymbol::create_extern_symbol(
-                "free",
-                CallingConvention::mock_arm32(),
-                Some(Datatype::Pointer),
-                None,
-            );
-            let other_function = ExternSymbol::create_extern_symbol(
-                "other_function",
-                CallingConvention::mock_arm32(),
-                None,
-                None,
-            );
-
-            Program::add_extern_symbols_to_program(vec![
-                (malloc.tid.clone(), malloc),
-                (free.tid.clone(), free),
-                (other_function.tid.clone(), other_function),
-            ])
-        }
-    }
-}
