+++ src/cwe_checker_lib/src/intermediate_representation/project.rs
@@ -291,75 +291,6 @@ impl Term<Jmp> {
 mod tests {
     use super::*;
 
-    impl Project {
-        /// Returns project with x64 calling convention and mocked program.
-        pub fn mock_x64() -> Project {
-            let mut none_cconv_register: Vec<Variable> = vec![
-                "RAX", "RBX", "RSP", "RBP", "R10", "R11", "R12", "R13", "R14", "R15",
-            ]
-            .into_iter()
-            .map(|name| Variable::mock(name, ByteSize::new(8)))
-            .collect();
-            let mut integer_register = CallingConvention::mock_x64().integer_parameter_register;
-            integer_register.append(&mut none_cconv_register);
-
-            let calling_conventions: BTreeMap<String, CallingConvention> =
-                BTreeMap::from([("__stdcall".to_string(), CallingConvention::mock_x64())]);
-
-            Project {
-                program: Term {
-                    tid: Tid::new("program_tid"),
-                    term: Program::mock_x64(),
-                },
-                cpu_architecture: "x86_64".to_string(),
-                stack_pointer_register: Variable::mock("RSP", 8u64),
-                calling_conventions,
-                register_set: integer_register.iter().cloned().collect(),
-                datatype_properties: DatatypeProperties::mock_x64(),
-                runtime_memory_image: RuntimeMemoryImage::mock(),
-            }
-        }
-
-        pub fn mock_arm32() -> Project {
-            let none_cconv_4byte_register: Vec<Variable> = vec!["r12", "r14", "r15"]
-                .into_iter()
-                .map(|name| Variable::mock(name, ByteSize::new(4)))
-                .collect();
-
-            let none_cconv_16byte_register: Vec<Variable> = vec![
-                "q0", "q1", "q2", "q3", "q8", "q9", "q10", "q11", "q12", "q13", "q14", "q15",
-            ]
-            .into_iter()
-            .map(|name| Variable::mock(name, ByteSize::new(16)))
-            .collect();
-
-            let callee_saved_register = CallingConvention::mock_arm32().callee_saved_register;
-
-            let integer_register = CallingConvention::mock_arm32()
-                .integer_parameter_register
-                .into_iter()
-                .chain(none_cconv_4byte_register)
-                .chain(none_cconv_16byte_register)
-                .chain(callee_saved_register);
-
-            Project {
-                program: Term {
-                    tid: Tid::new("program_tid"),
-                    term: Program::mock_arm32(),
-                },
-                cpu_architecture: "arm32".to_string(),
-                stack_pointer_register: Variable::mock("sp", 4u64),
-                calling_conventions: BTreeMap::from([(
-                    "__stdcall".to_string(),
-                    CallingConvention::mock_arm32(),
-                )]),
-                register_set: integer_register.collect(),
-                datatype_properties: DatatypeProperties::mock_arm32(),
-                runtime_memory_image: RuntimeMemoryImage::mock(),
-            }
-        }
-    }
-
     #[test]
     fn retarget_nonexisting_jumps() {
         let mut jmp_term = Term {
