+++ src/cwe_checker_lib/src/intermediate_representation/sub.rs
@@ -195,232 +195,3 @@ impl CallingConvention {
         register_list
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    impl Sub {
-        pub fn mock(name: impl ToString) -> Term<Sub> {
-            Term {
-                tid: Tid::new(name.to_string()),
-                term: Sub {
-                    name: name.to_string(),
-                    blocks: Vec::new(),
-                    calling_convention: None,
-                },
-            }
-        }
-    }
-    /// Wrapper for subpiece to model float register for argument passing
-    fn create_float_register_subpiece(
-        name: &str,
-        reg_size: u64,
-        low_byte: u64,
-        size: u64,
-    ) -> Expression {
-        Expression::subpiece(
-            Expression::Var(Variable::mock(name, reg_size)),
-            ByteSize::new(low_byte),
-            ByteSize::new(size),
-        )
-    }
-
-    impl CallingConvention {
-        /// Creates System V Calling Convention with Advanced Vector Extensions 512
-        pub fn mock_x64() -> CallingConvention {
-            CallingConvention {
-                name: "__stdcall".to_string(), // so that the mock is useable as standard calling convention in tests
-                integer_parameter_register: vec![
-                    Variable::mock("RDI", 8),
-                    Variable::mock("RSI", 8),
-                    Variable::mock("RDX", 8),
-                    Variable::mock("RCX", 8),
-                    Variable::mock("R8", 8),
-                    Variable::mock("R9", 8),
-                ],
-                // ABI: first 8 Bytes of ZMM0-ZMM7 for float parameter
-                // Ghidra: first 8 Bytes of YMM0-YMM7 for float parameter
-                float_parameter_register: vec![
-                    create_float_register_subpiece("ZMM0", 64, 0, 8),
-                    create_float_register_subpiece("ZMM1", 64, 0, 8),
-                    create_float_register_subpiece("ZMM2", 64, 0, 8),
-                    create_float_register_subpiece("ZMM3", 64, 0, 8),
-                    create_float_register_subpiece("ZMM4", 64, 0, 8),
-                    create_float_register_subpiece("ZMM5", 64, 0, 8),
-                    create_float_register_subpiece("ZMM6", 64, 0, 8),
-                    create_float_register_subpiece("ZMM7", 64, 0, 8),
-                ],
-                integer_return_register: vec![Variable::mock("RAX", 8), Variable::mock("RDX", 8)],
-                // ABI: XMM0-XMM1 float return register
-                // Ghidra: uses XMM0 only
-                float_return_register: vec![create_float_register_subpiece("ZMM0", 64, 0, 8)],
-                callee_saved_register: vec![
-                    Variable::mock("RBP", 8),
-                    Variable::mock("RBX", 8),
-                    Variable::mock("RSP", 8),
-                    Variable::mock("R12", 8),
-                    Variable::mock("R13", 8),
-                    Variable::mock("R14", 8),
-                    Variable::mock("R15", 8),
-                ],
-            }
-        }
-        /// Following ARM32 ABI with MVE Extention
-        pub fn mock_arm32() -> CallingConvention {
-            CallingConvention {
-                name: "__stdcall".to_string(), // so that the mock is useable as standard calling convention in tests
-                integer_parameter_register: vec![
-                    Variable::mock("r0", 4),
-                    Variable::mock("r1", 4),
-                    Variable::mock("r2", 4),
-                    Variable::mock("r3", 4),
-                ],
-                // ABI: q0-q3 used for argument passing
-                // Ghidra: uses q0-q1 only
-                float_parameter_register: vec![
-                    create_float_register_subpiece("q0", 16, 0, 4),
-                    create_float_register_subpiece("q0", 16, 4, 4),
-                    create_float_register_subpiece("q0", 16, 8, 4),
-                    create_float_register_subpiece("q0", 16, 12, 4),
-                    create_float_register_subpiece("q1", 16, 0, 4),
-                    create_float_register_subpiece("q1", 16, 4, 4),
-                    create_float_register_subpiece("q1", 16, 8, 4),
-                    create_float_register_subpiece("q1", 16, 12, 4),
-                ],
-                // ABI: r0-r1 used as integer return register
-                // Ghidra uses r0 only
-                integer_return_register: vec![
-                    Variable::mock("r0", 4),
-                    Variable::mock("r1", 4),
-                    Variable::mock("r2", 4),
-                    Variable::mock("r3", 4),
-                ],
-                // ABI: whole q0 used as float return
-                // Ghidra: uses first 8 Bytes of q0 only
-                float_return_register: vec![create_float_register_subpiece("q0", 16, 0, 4)],
-                callee_saved_register: vec![
-                    Variable::mock("r4", 4),
-                    Variable::mock("r5", 4),
-                    Variable::mock("r6", 4),
-                    Variable::mock("r7", 4),
-                    Variable::mock("r8", 4),
-                    Variable::mock("r9", 4),
-                    Variable::mock("r10", 4),
-                    Variable::mock("r11", 4),
-                    Variable::mock("r13", 4),
-                    Variable::mock("q4", 16),
-                    Variable::mock("q5", 16),
-                    Variable::mock("q6", 16),
-                    Variable::mock("q7", 16),
-                ],
-            }
-        }
-    }
-
-    impl Arg {
-        pub fn mock_register(name: impl ToString, size_in_bytes: impl Into<ByteSize>) -> Arg {
-            Arg::Register {
-                expr: Expression::Var(Variable::mock(name.to_string(), size_in_bytes)),
-                data_type: None,
-            }
-        }
-
-        pub fn mock_register_with_data_type(
-            name: impl ToString,
-            size_in_bytes: impl Into<ByteSize>,
-            data_type: Option<Datatype>,
-        ) -> Arg {
-            Arg::Register {
-                expr: Expression::Var(Variable::mock(name.to_string(), size_in_bytes)),
-                data_type,
-            }
-        }
-
-        pub fn mock_pointer_register(
-            name: impl ToString,
-            size_in_bytes: impl Into<ByteSize>,
-        ) -> Arg {
-            Arg::Register {
-                expr: Expression::Var(Variable::mock(name.to_string(), size_in_bytes)),
-                data_type: Some(Datatype::Pointer),
-            }
-        }
-    }
-
-    impl ExternSymbol {
-        pub fn mock_x64(name: impl ToString) -> ExternSymbol {
-            ExternSymbol {
-                tid: Tid::new(name.to_string()),
-                addresses: vec!["UNKNOWN".to_string()],
-                name: name.to_string(),
-                calling_convention: Some("__stdcall".to_string()),
-                parameters: vec![Arg::mock_register("RDI", 8)],
-                return_values: vec![Arg::mock_register("RAX", 8)],
-                no_return: false,
-                has_var_args: false,
-            }
-        }
-
-        pub fn mock_arm32(name: impl ToString) -> ExternSymbol {
-            ExternSymbol {
-                tid: Tid::new(name.to_string()),
-                addresses: vec!["UNKNOWN".to_string()],
-                name: name.to_string(),
-                calling_convention: Some("__stdcall".to_string()),
-                parameters: vec![Arg::mock_register("r0", 4)],
-                return_values: vec![Arg::mock_register("r0", 4)],
-                no_return: false,
-                has_var_args: false,
-            }
-        }
-
-        pub fn mock_sprintf_x64() -> Self {
-            ExternSymbol {
-                tid: Tid::new("sprintf"),
-                addresses: vec!["UNKNOWN".to_string()],
-                name: "sprintf".to_string(),
-                calling_convention: Some("__stdcall".to_string()),
-                parameters: vec![Arg::mock_register("RDI", 8), Arg::mock_register("RSI", 8)],
-                return_values: vec![Arg::mock_register("RAX", 8)],
-                no_return: false,
-                has_var_args: true,
-            }
-        }
-
-        /// Returns extern symbol with argument/return register according to calling convention
-        pub fn create_extern_symbol(
-            name: &str,
-            cconv: CallingConvention,
-            arg_type: Option<Datatype>,
-            return_type: Option<Datatype>,
-        ) -> ExternSymbol {
-            ExternSymbol {
-                tid: Tid::new(name),
-                addresses: vec![],
-                name: name.to_string(),
-                calling_convention: Some(cconv.name),
-                parameters: match arg_type {
-                    Some(data_type) => {
-                        vec![Arg::from_var(
-                            cconv.integer_parameter_register[0].clone(),
-                            Some(data_type),
-                        )]
-                    }
-                    None => vec![],
-                },
-                return_values: match return_type {
-                    Some(data_type) => {
-                        vec![Arg::from_var(
-                            cconv.integer_return_register[0].clone(),
-                            Some(data_type),
-                        )]
-                    }
-                    None => vec![],
-                },
-                no_return: false,
-                has_var_args: false,
-            }
-        }
-    }
-}
