+++ src/cwe_checker_lib/src/intermediate_representation/variable.rs
@@ -29,18 +29,3 @@ impl Display for Variable {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    impl Variable {
-        pub fn mock(name: impl ToString, size_in_bytes: impl Into<ByteSize>) -> Variable {
-            Variable {
-                name: name.to_string(),
-                size: size_in_bytes.into(),
-                is_temp: false,
-            }
-        }
-    }
-}
